ddl: migrate test-infra to testify for ddl.testDBSuite4 (#33113)
close pingcap/tidb#32964
